#203 Display Avatar of users assigned to a task
